requirements.txt
airsim>=1.8.0
numpy>=1.21.0
opencv-python>=4.5.0
msgpack-rpc-python